[WIP] Add GPU File Manger

+++ libgeminifs/include/utils.cuh
@@ -66,7 +66,7 @@ static __device__ void cuda_device_assert(cudaError_t code, const char *file, in
 using cuda_device_lock = cuda::binary_semaphore<cuda::thread_scope_device>;
 using cuda_device_ref = cuda::atomic<uint32_t, cuda::thread_scope_device>;
 using FileGroupId = int32_t;
-using GPUFileId = uint64_t;
+using GPUFileId = uint32_t;
